made health bar resilient and functional, added fall damage (uses collider trigger and if velocity is massive, might not register. seems fine for practical application at the moment though

+++ Assets/Scripts/Enemy/UnitHealthBar.cs
@@ -1,27 +1,27 @@
 ﻿using Assets.Scripts;
+using System;
 using UnityEngine;
 
 public class UnitHealthBar : MonoBehaviour
 {
-	private Unit parentUnit;
+    private Unit parentUnit;
 
-	// Use this for initialization
-	void Start()
-	{
-		parentUnit = GetComponentInParent<Unit>();
-	}
+    void Start()
+    {
+        parentUnit = GetComponentInParent<Unit>();
+    }
 
-	// Update is called once per frame
-	void Update()
-	{
-		transform.localScale = new Vector3(parentUnit.HealthCurrent / parentUnit.HealthMax, transform.localScale.y, transform.localScale.z);
+    void Update()
+    {
+        //cast is NOT redundant
+        transform.localScale = new Vector3((((float)parentUnit.HealthCurrent) / (float)parentUnit.HealthMax), transform.localScale.y, transform.localScale.z);
 
-		////flip the same way Unit does, so when unit is flipped, this is flipped back to original direction
-		//var scale = new Vector3(Mathf.Abs(transform.localScale.x), transform.localScale.y, transform.localScale.z);
-		//if (parentUnit.UnitState.HasFlag(Enums.UnitStateEnum.FacingRight))
-		//{
-		//	scale.x *= -1;
-		//}
-		//transform.localScale = scale;
-	}
+        ////flip the same way Unit does, so when unit is flipped, this is flipped back to original direction
+        //var scale = new Vector3(Mathf.Abs(transform.localScale.x), transform.localScale.y, transform.localScale.z);
+        //if (parentUnit.UnitState.HasFlag(Enums.UnitStateEnum.FacingRight))
+        //{
+        //	scale.x *= -1;
+        //}
+        //transform.localScale = scale;
+    }
 }
